Refactor routing. Create auth and api utils. Add User and EventBoard contexts
