+++ .env
@@ -1,3 +1,5 @@
+# private
 SKIP_PREFLIGHT_CHECK=true
-ACCESS_TOKEN_SECRET=67e3cddf8b0254b3a13288ff2d9d0fec8fd0a93f242e7e01bc2ead8f66c8dae42c80bb80cfdbbbcdfef243904e42537da9c495516690ce0669ad7db1d29e7fcd
-REFRESH_TOKEN_SECRET=70c52e16d3b136ffa904f3f8f4f8b8e44394985828b0bced2945efdf392222655f4079db39432134e37870e399a5c5d76fd0d58492d55da1baa4154eb2dbb052
+
+# public
+REACT_APP_API_URL=http://localhost:8002/api
