+++ package.json
@@ -5,7 +5,6 @@
   "dependencies": {
     "@babel/code-frame": "7.14.5",
     "@testing-library/react-hooks": "7.0.1",
-    "axios": "0.21.1",
     "babel": "6.23.0",
     "classnames": "2.3.1",
     "cors": "2.8.5",
