+++ src/index.js
@@ -3,6 +3,6 @@ import ReactDOM from "react-dom";
 
 import "./index.scss";
 
-import Application from "./components/Application.js";
+import Application from "./components/Application";
 
 ReactDOM.render(<Application />, document.getElementById("root"));
